Maintenance
Pop up when the user is stupid
Refactoring merge function

+++ szakdolgozat/szakdoga/objrepair.h
@@ -59,7 +59,7 @@ namespace approx{
 
 		//elemszam
 		int size() const {
-			return (int)ind_map.size();
+			return ind_map.size();
 		}
 
 		ConstIterator begin() const { return ConstIterator(&vecs,&ind_map,0); }
@@ -119,7 +119,7 @@ namespace approx{
 			inds.push_back(vmap[v]);
 		}
 		int transform_index(int ind) const { return inds[ind]; }
-		int size()const { return (int)vecs.size(); }
+		int size()const { return vecs.size(); }
 		Vector3<T> operator[](int ind) const {
 			return vecs[inds[ind]];
 		}
